restaurant-service/requirements.txt
flask==3.1.0
flask_sqlalchemy==3.1.1
flask_cors==5.0.0
psycopg2==2.9.10
redis==5.2.0
python-dotenv
psycopg2-binary==2.9.10
prometheus_client==0.16.0
